pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel", "setuptools-rust>=1.7.0"]
build-backend = "setuptools.build_meta"

[project]
name = "torpedocode"
version = "0.1.0"
description = "Hybrid topological + neural framework for limit order book modelling"
authors = [{name = "TorpedoCode Research"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=1.5",
    "torch>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "maturin>=1.5",
    "ripser>=0.6; platform_system != 'Windows'",
    "persim>=0.3",
    "gudhi>=3.9; platform_system != 'Windows'",
]

[tool.setuptools]
package-dir = {"" = "python"}

[tool.setuptools.packages.find]
where = ["python"]

[tool.setuptools-rust]
binding = "pyo3"
strip = true
debug = false

[[tool.setuptools-rust.ext-modules]]
target = "torpedocode_ingest"
path = "rust/torpedocode-py/Cargo.toml"
optional = true

[[tool.setuptools-rust.ext-modules]]
target = "torpedocode_tda"
path = "rust/torpedocode-tda/Cargo.toml"
optional = true

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
